Move EndpointProvider under data-model-providers/codedriven/endpoint
