+++ src/data-model-providers/codedriven/endpoint/tests/BUILD.gn
@@ -28,6 +28,6 @@ chip_test_suite("tests") {
     "${chip_root}/src/app/clusters/descriptor",
     "${chip_root}/src/app/data-model-provider:string-builder-adapters",
     "${chip_root}/src/app/server-cluster/testing",
-    "${chip_root}/src/data-model-providers/endpoint",
+    "${chip_root}/src/data-model-providers/codedriven/endpoint",
   ]
 }
